# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "email": "Jacksontpa7@gmail.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/bills/create-bill": {
            "post": {
                "description": "Create a bill with participants and their items",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bill"
                ],
                "summary": "Create a new bill",
                "parameters": [
                    {
                        "description": "Bill Info",
                        "name": "bill",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateBillRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateBillResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/check": {
            "get": {
                "description": "Check backend health",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Check backend health",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/confirm-payment": {
            "get": {
                "description": "example return confirms a payment on Hedera",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Example confirm a payment",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/participants/get-all-participant-detail/{participantId}": {
            "get": {
                "description": "Get participant all bills",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Participants"
                ],
                "summary": "Get participant Bills",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Participant ID",
                        "name": "participantId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.ParticipantDetailResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/participants/get-participant-detail": {
            "post": {
                "description": "Retrieve participant detail including bill and items using billId and participantId",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Participant"
                ],
                "summary": "Get participant detail in a bill",
                "parameters": [
                    {
                        "description": "Bill and Participant IDs",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.GetParticipantDetailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.ParticipantDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users/register": {
            "post": {
                "description": "Save wallet address to database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User info",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dtos.CreateBillItemResponse": {
            "type": "object",
            "properties": {
                "itemId": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                }
            }
        },
        "dtos.CreateBillParticipantResponse": {
            "type": "object",
            "properties": {
                "amountOwed": {
                    "type": "integer"
                },
                "isPaid": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.CreateBillItemResponse"
                    }
                },
                "participantId": {
                    "type": "string"
                }
            }
        },
        "dtos.CreateBillRequest": {
            "type": "object",
            "properties": {
                "billTitle": {
                    "type": "string",
                    "example": "Dinner at Cafe"
                },
                "creatorId": {
                    "type": "string",
                    "example": "user123"
                },
                "participants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.CreateParticipantRequest"
                    }
                },
                "totalAmount": {
                    "type": "integer",
                    "example": 200000
                }
            }
        },
        "dtos.CreateBillResponse": {
            "type": "object",
            "properties": {
                "billId": {
                    "type": "string"
                },
                "billTitle": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "creatorId": {
                    "type": "string"
                },
                "participants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.CreateBillParticipantResponse"
                    }
                },
                "totalAmount": {
                    "type": "integer"
                }
            }
        },
        "dtos.CreateItemRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Steak"
                },
                "price": {
                    "type": "integer",
                    "example": 80000
                }
            }
        },
        "dtos.CreateParticipantRequest": {
            "type": "object",
            "properties": {
                "amountOwed": {
                    "type": "integer",
                    "example": 100000
                },
                "isPaid": {
                    "type": "boolean",
                    "example": true
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.CreateItemRequest"
                    }
                },
                "participantId": {
                    "type": "string",
                    "example": "user123"
                }
            }
        },
        "dtos.GetParticipantDetailRequest": {
            "type": "object",
            "required": [
                "billId",
                "participantId"
            ],
            "properties": {
                "billId": {
                    "type": "string"
                },
                "participantId": {
                    "type": "string"
                }
            }
        },
        "dtos.ParticipantDetailResponse": {
            "type": "object",
            "properties": {
                "billId": {
                    "type": "string"
                },
                "billTitle": {
                    "type": "string"
                },
                "creatorId": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Item"
                    }
                },
                "participantId": {
                    "type": "string"
                },
                "totalOwed": {
                    "type": "integer"
                }
            }
        },
        "models.Item": {
            "type": "object",
            "properties": {
                "billId": {
                    "type": "string"
                },
                "itemId": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "participantId": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "wallet_address": {
                    "type": "string"
                }
            }
        }
    }
}